third_party/blink/web_tests/wpt_internal/webgpu/000_run_me_first.https.html

<!--
  TODO(crbug.com/953991): The first WebGPU test gets blocked on some browser
  startup work that takes a long time, causing the first test to flake. This
  "test" just exists to sit at the beginning of the test list and be marked as
  [ Slow ].
-->
<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  'use strict';

  promise_test(async t => {
    const adapter = await navigator.gpu.requestAdapter();
    assert_true(adapter !== null);
    const device = await adapter.requestDevice();
    assert_true(device !== null);
  });
</script>
